Created Profile Component

+++ mysite/app/Http/Controllers/posts/PostsController.php
@@ -8,7 +8,10 @@ use App\Models\Post;
 
 class PostsController extends Controller
 {
-    //
+    public function __construct()
+    {
+        $this->middleware("auth");
+    }
     public function index()
     {
         $posts = Post::latest()->with("user")->paginate(5);
@@ -18,6 +21,7 @@ class PostsController extends Controller
     }
     public function delete(Post $post)
     {
+        $this->authorize("authorized", $post);
         $post->delete();
         return redirect()->route("home");
     }
